+++ src/features/trading/containers/OrderContainer.tsx
@@ -1,541 +1,11 @@
 "use client";
 
-import React, { useState, useEffect, useRef, useCallback } from "react";
-import OrderForm from "@/features/trading/components/OrderForm";
-import { useMarketPrice } from "@/features/trading/hooks/useMarketPrice";
+import React from "react";
 import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs";
-import { useSession } from "next-auth/react";
-import { useRouter } from "next/navigation";
-import { useGetCashBalance } from "@/features/wallet/hooks/useGetCash";
-import { useCreateBuyOrder } from "@/features/trading/hooks/useCreateBuyOrder";
-import { useCreateSellOrder } from "@/features/trading/hooks/useCreateSellOrder";
-import { useQueryClient } from "@tanstack/react-query";
-import { TradeQueryKeys } from "@/features/trading/constants";
-import { useGetCoin } from "@/features/trading/hooks/useGetCoin";
+import BuyOrderContainer from "@/features/trading/containers/BuyOrderContainer";
+import SellOrderContainer from "@/features/trading/containers/SellOrderContainer";
 
 export default function MarketOrderContainer() {
-  const inputRef = useRef<HTMLInputElement>(null);
-  const amountInputRef = useRef<HTMLInputElement>(null);
-  const marketPrice = useMarketPrice();
-  const { data: session } = useSession();
-  const router = useRouter();
-  const queryClient = useQueryClient();
-
-  // Fetch wallet balance
-  const {
-    data: cashBalance,
-    isLoading: isBalanceLoading,
-    error: balanceError,
-  } = useGetCashBalance({
-    enabled: !!session,
-  });
-
-  // Fetch BTC balance
-  const {
-    data: btcBalance,
-    isLoading: isBtcBalanceLoading,
-    error: btcBalanceError,
-  } = useGetCoin({
-    symbol: "BTC",
-    enabled: !!session,
-  });
-
-  // Create buy order mutation
-  const createBuyOrderMutation = useCreateBuyOrder({
-    onSuccess: (data) => {
-      console.log("Buy order created successfully:", data);
-      queryClient.invalidateQueries({
-        queryKey: [TradeQueryKeys.GET_CASH_BALANCE],
-      });
-      queryClient.invalidateQueries({
-        queryKey: [TradeQueryKeys.GET_COIN_ASSET, "BTC"],
-      });
-
-      if (data.filled > 0) {
-        const filledUSD =
-          data.spent || data.filled * parseFloat(price.replace(/,/g, ""));
-        alert(
-          `Buy BTC/USDT Amount ${filledUSD.toFixed(
-            2
-          )} USD submitted successfully`
-        );
-      } else if (data.remaining > 0 && data.filled === 0) {
-        alert(
-          `📝 Order created successfully!\n` +
-            `Order ID: ${data.orderRef}\n` +
-            `Amount remaining: ${data.remaining.toFixed(8)} BTC\n` +
-            `Status: Pending`
-        );
-      } else {
-        let message = `Order ID: ${data.orderRef}`;
-        if (data.refund > 0) {
-          const actualSpent =
-            parseFloat(amount.replace(/,/g, "")) - data.refund;
-          message += `\nActual spent: ${actualSpent.toFixed(2)} USD`;
-          message += `\nRefund: ${data.refund.toFixed(2)} USD`;
-          if (data.message) {
-            message += `\nMessage: ${data.message}`;
-          }
-        }
-        alert(message);
-      }
-      handleSubmitSuccess("buy");
-    },
-  });
-
-  // Create sell order mutation
-  const createSellOrderMutation = useCreateSellOrder({
-    onSuccess: (data) => {
-      console.log("Sell order created successfully:", data);
-      queryClient.invalidateQueries({
-        queryKey: [TradeQueryKeys.GET_CASH_BALANCE],
-      });
-      queryClient.invalidateQueries({
-        queryKey: [TradeQueryKeys.GET_COIN_ASSET, "BTC"],
-      });
-
-      // Show success message with updated response format
-      if (data.filled > 0) {
-        alert(
-          `✅ Sell order completed successfully!\n` +
-            `Order ID: ${data.orderRef}\n` +
-            `BTC Sold: ${data.filled.toFixed(8)}\n` +
-            `Proceeds: $${data.proceeds.toFixed(2)}`
-        );
-      } else {
-        alert(
-          `📝 Sell order created successfully!\n` +
-            `Order ID: ${data.orderRef}\n` +
-            `Status: Pending`
-        );
-      }
-      handleSubmitSuccess("sell");
-    },
-    onError: (error) => {
-      console.error("Sell order error:", error);
-      alert(`Error creating sell order: ${error.message}`);
-    },
-  });
-
-  // Common states
-  const [priceLabel, setPriceLabel] = useState("Price");
-  const [isInputFocused, setIsInputFocused] = useState(false);
-  const [limitPrice, setLimitPrice] = useState<string>("");
-  const [price, setPrice] = useState<string>(marketPrice);
-
-  // Buy states
-  const [amount, setAmount] = useState<string>("");
-  const [isAmountFocused, setIsAmountFocused] = useState(false);
-  const [isAmountValid, setIsAmountValid] = useState(true);
-  const [sliderValue, setSliderValue] = useState<number>(0);
-  const [receiveBTC, setReceiveBTC] = useState<string>("");
-
-  // Sell states
-  const [sellAmount, setSellAmount] = useState<string>("");
-  const [isSellAmountValid, setIsSellAmountValid] = useState(true);
-  const [sellSliderValue, setSellSliderValue] = useState<number>(0);
-  const [receiveUSD, setReceiveUSD] = useState<string>("");
-  const [isSellAmountFocused, setIsSellAmountFocused] = useState(false);
-
-  // Get available balance dynamically for USD
-  const getAvailableBalance = useCallback(() => {
-    if (!session || !cashBalance) return 0;
-    return cashBalance.amount || 0;
-  }, [session, cashBalance]);
-
-  // Get available BTC balance dynamically
-  const getAvailableBTCBalance = useCallback(() => {
-    if (!session || !btcBalance) return 0;
-    return btcBalance.amount || 0;
-  }, [session, btcBalance]);
-
-  // Format available balance for display
-  const formatAvailableBalance = useCallback(() => {
-    const balance = getAvailableBalance();
-    return new Intl.NumberFormat("en-US", {
-      minimumFractionDigits: 2,
-      maximumFractionDigits: 2,
-    }).format(balance);
-  }, [getAvailableBalance]);
-
-  // Format available BTC balance for display
-  const formatAvailableBTCBalance = useCallback(() => {
-    const balance = getAvailableBTCBalance();
-    return balance.toFixed(9); // แสดงทศนิยม 9 ตำแหน่งสำหรับ BTC
-  }, [getAvailableBTCBalance]);
-
-  // Helper functions
-  const formatNumberWithComma = useCallback((value: string): string => {
-    if (!value) return "";
-    const numericValue = value.replace(/,/g, "");
-    if (!/^\d*\.?\d*$/.test(numericValue)) return value;
-    const parts = numericValue.split(".");
-    const integerPart = parts[0];
-    const decimalPart = parts[1];
-    const formattedInteger = integerPart.replace(/\B(?=(\d{3})+(?!\d))/g, ",");
-    return decimalPart !== undefined
-      ? `${formattedInteger}.${decimalPart}`
-      : formattedInteger;
-  }, []);
-
-  const formatToTwoDecimalsWithComma = useCallback(
-    (value: string): string => {
-      if (!value) return "";
-      const numericValue = value.replace(/,/g, "");
-      const num = parseFloat(numericValue);
-      if (isNaN(num)) return "";
-      return formatNumberWithComma(num.toFixed(2));
-    },
-    [formatNumberWithComma]
-  );
-
-  const formatBTCNumber = useCallback((value: string): string => {
-    if (!value) return "";
-    const numericValue = value.replace(/,/g, "");
-    if (!/^\d*\.?\d{0,9}$/.test(numericValue)) return value;
-    return numericValue;
-  }, []);
-
-  const isValidNumberFormat = useCallback((value: string): boolean => {
-    const numericValue = value.replace(/,/g, "");
-    return /^\d*\.?\d{0,2}$/.test(numericValue);
-  }, []);
-
-  const isValidBTCFormat = useCallback((value: string): boolean => {
-    const numericValue = value.replace(/,/g, "");
-    return /^\d*\.?\d{0,9}$/.test(numericValue);
-  }, []);
-
-  const calculateReceiveBTC = useCallback(
-    (amountValue: string, priceValue: string): string => {
-      if (!amountValue || !priceValue) return "";
-      const numAmount = parseFloat(amountValue.replace(/,/g, ""));
-      const numPrice = parseFloat(priceValue.replace(/,/g, ""));
-      if (isNaN(numAmount) || isNaN(numPrice) || numPrice <= 0) return "";
-      const btcAmount = numAmount / numPrice;
-      return btcAmount.toFixed(9);
-    },
-    []
-  );
-
-  const calculateReceiveUSD = useCallback(
-    (btcAmount: string, priceValue: string): string => {
-      if (!btcAmount || !priceValue) return "";
-      const numBTC = parseFloat(btcAmount);
-      const numPrice = parseFloat(priceValue.replace(/,/g, ""));
-      if (isNaN(numBTC) || isNaN(numPrice) || numPrice <= 0) return "";
-      const usdAmount = numBTC * numPrice;
-      return formatNumberWithComma(usdAmount.toFixed(2));
-    },
-    [formatNumberWithComma]
-  );
-
-  const calculateSliderPercentage = useCallback(
-    (amountValue: string): number => {
-      if (!amountValue) return 0;
-      const numAmount = parseFloat(amountValue.replace(/,/g, ""));
-      const availableBalance = getAvailableBalance();
-      if (isNaN(numAmount) || numAmount <= 0 || availableBalance <= 0) return 0;
-      const percentage = (numAmount / availableBalance) * 100;
-      return Math.min(percentage, 100);
-    },
-    [getAvailableBalance]
-  );
-
-  const calculateSellSliderPercentage = useCallback(
-    (btcAmount: string): number => {
-      if (!btcAmount) return 0;
-      const numAmount = parseFloat(btcAmount);
-      const availableBTC = getAvailableBTCBalance();
-      if (isNaN(numAmount) || numAmount <= 0 || availableBTC <= 0) return 0;
-      const percentage = (numAmount / availableBTC) * 100;
-      return Math.min(percentage, 100);
-    },
-    [getAvailableBTCBalance]
-  );
-
-  const calculateAmountFromPercentage = useCallback(
-    (percentage: number): string => {
-      const availableBalance = getAvailableBalance();
-      const amount = (percentage / 100) * availableBalance;
-      return formatNumberWithComma(amount.toFixed(2));
-    },
-    [getAvailableBalance, formatNumberWithComma]
-  );
-
-  const calculateBTCFromPercentage = useCallback(
-    (percentage: number): string => {
-      const availableBTC = getAvailableBTCBalance();
-      const btcAmount = (percentage / 100) * availableBTC;
-      return btcAmount.toFixed(9);
-    },
-    [getAvailableBTCBalance]
-  );
-
-  // Price handlers
-  const handlePriceFocus = () => {
-    setPriceLabel("Limit price");
-    setIsInputFocused(true);
-    setPrice("");
-    setLimitPrice("");
-  };
-
-  const handleMarketClick = () => {
-    setPriceLabel("Price");
-    const formattedMarketPrice = formatToTwoDecimalsWithComma(marketPrice);
-    setPrice(formattedMarketPrice);
-    setLimitPrice(formattedMarketPrice);
-    setIsInputFocused(false);
-  };
-
-  const handlePriceChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const inputValue = e.target.value;
-    if (inputValue === "" || isValidNumberFormat(inputValue)) {
-      const formattedValue = formatNumberWithComma(inputValue);
-      setLimitPrice(formattedValue);
-      setPrice(formattedValue);
-    }
-  };
-
-  const handlePriceBlur = () => {
-    if (price) {
-      const formattedPrice = formatToTwoDecimalsWithComma(price);
-      setPrice(formattedPrice);
-      setLimitPrice(formattedPrice);
-    } else if (priceLabel === "Price" && marketPrice) {
-      const formattedMarketPrice = formatToTwoDecimalsWithComma(marketPrice);
-      setPrice(formattedMarketPrice);
-      setLimitPrice(formattedMarketPrice);
-    }
-    setIsInputFocused(false);
-  };
-
-  // Buy handlers
-  const handleAmountChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const inputValue = e.target.value;
-    if (inputValue === "" || isValidNumberFormat(inputValue)) {
-      const formattedValue = formatNumberWithComma(inputValue);
-      setAmount(formattedValue);
-
-      const numericValue = inputValue.replace(/,/g, "");
-      const num = parseFloat(numericValue);
-      const availableBalance = getAvailableBalance();
-      const isValid =
-        inputValue === "" || (!isNaN(num) && num <= availableBalance);
-
-      if (!isValid && inputValue !== "") {
-        setSliderValue(0);
-      } else {
-        const sliderPercentage = calculateSliderPercentage(inputValue);
-        setSliderValue(sliderPercentage);
-      }
-      setIsAmountValid(isValid);
-    }
-  };
-
-  const handleSliderChange = (percentage: number) => {
-    setSliderValue(percentage);
-    const newAmount = calculateAmountFromPercentage(percentage);
-    setAmount(newAmount);
-    const numericValue = newAmount.replace(/,/g, "");
-    const num = parseFloat(numericValue);
-    const availableBalance = getAvailableBalance();
-    setIsAmountValid(!isNaN(num) && num <= availableBalance);
-  };
-
-  const handleAmountFocus = () => setIsAmountFocused(true);
-
-  const handleAmountBlur = () => {
-    setIsAmountFocused(false);
-    if (amount) {
-      const numericValue = amount.replace(/,/g, "");
-      const num = parseFloat(numericValue);
-      if (!isNaN(num)) {
-        const formattedAmount = formatNumberWithComma(num.toFixed(2));
-        setAmount(formattedAmount);
-      }
-    }
-  };
-
-  // Sell handlers
-  const handleSellAmountChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const inputValue = e.target.value;
-    if (inputValue === "" || isValidBTCFormat(inputValue)) {
-      const formattedValue = formatBTCNumber(inputValue);
-      setSellAmount(formattedValue);
-
-      const num = parseFloat(inputValue);
-      const availableBTC = getAvailableBTCBalance();
-      const isValid = inputValue === "" || (!isNaN(num) && num <= availableBTC);
-
-      if (!isValid && inputValue !== "") {
-        setSellSliderValue(0);
-      } else {
-        const sliderPercentage = calculateSellSliderPercentage(inputValue);
-        setSellSliderValue(sliderPercentage);
-      }
-      setIsSellAmountValid(isValid);
-    }
-  };
-
-  const handleSellSliderChange = (percentage: number) => {
-    setSellSliderValue(percentage);
-    const newBTCAmount = calculateBTCFromPercentage(percentage);
-    setSellAmount(newBTCAmount);
-    const num = parseFloat(newBTCAmount);
-    const availableBTC = getAvailableBTCBalance();
-    setIsSellAmountValid(!isNaN(num) && num <= availableBTC);
-  };
-
-  const handleSellAmountFocus = () => setIsSellAmountFocused(true);
-
-  const handleSellAmountBlur = () => {
-    setIsSellAmountFocused(false);
-    if (sellAmount) {
-      const num = parseFloat(sellAmount);
-      if (!isNaN(num)) setSellAmount(num.toFixed(9));
-    }
-  };
-
-  // Submit handler
-  const handleSubmit = (type: "buy" | "sell") => {
-    if (!session) {
-      alert("Please login to continue trading");
-      router.push("/auth/sign-in");
-      return;
-    }
-
-    const amountToSubmit = type === "buy" ? amount : sellAmount;
-    const numericAmount = parseFloat(amountToSubmit.replace(/,/g, "") || "0");
-
-    if (!numericAmount || numericAmount === 0) {
-      return;
-    }
-
-    if (type === "buy") {
-      const numericPrice = parseFloat(price.replace(/,/g, "") || "0");
-      const btcAmount = parseFloat(receiveBTC.replace(/,/g, "") || "0");
-      const userId =
-        cashBalance?.userId || (session.user as any)?.id || session.user?.email;
-      const orderPayload = {
-        userId: userId,
-        symbol: "BTC",
-        price: numericPrice,
-        amount: btcAmount,
-      };
-
-      console.log("Buy order payload:", orderPayload);
-      console.log("USD to spend:", numericAmount);
-      console.log("Price per BTC:", numericPrice);
-      console.log("BTC amount to buy:", btcAmount);
-
-      createBuyOrderMutation.mutate(orderPayload);
-    } else {
-      // Handle sell order submission with updated payload
-      const numericPrice = parseFloat(price.replace(/,/g, "") || "0");
-      const btcAmountToSell = parseFloat(sellAmount || "0");
-      const userId =
-        cashBalance?.userId || (session.user as any)?.id || session.user?.email;
-
-      // Calculate lotPrice (total value of the trade)
-      const lotPrice = numericPrice * btcAmountToSell;
-
-      const sellOrderPayload = {
-        userId: userId,
-        symbol: "BTC",
-        price: numericPrice,
-        amount: btcAmountToSell,
-        lotPrice: lotPrice, // เพิ่ม lotPrice ตาม API spec
-      };
-
-      console.log("Sell order payload:", sellOrderPayload);
-      console.log("BTC amount to sell:", btcAmountToSell);
-      console.log("Price per BTC:", numericPrice);
-      console.log("Lot Price (total value):", lotPrice);
-      console.log("USD to receive:", receiveUSD);
-
-      createSellOrderMutation.mutate(sellOrderPayload);
-    }
-  };
-
-  const handleSubmitSuccess = (type: "buy" | "sell") => {
-    if (type === "buy") {
-      setAmount("");
-      setSliderValue(0);
-      setReceiveBTC("");
-    } else {
-      setSellAmount("");
-      setSellSliderValue(0);
-      setReceiveUSD("");
-    }
-  };
-
-  // Effects
-  useEffect(() => {
-    if (priceLabel === "Price") {
-      const formattedPrice = formatNumberWithComma(marketPrice);
-      setPrice(formattedPrice);
-    }
-  }, [marketPrice, priceLabel, formatNumberWithComma]);
-
-  useEffect(() => {
-    const currentPrice = priceLabel === "Price" ? marketPrice : limitPrice;
-    const btcAmount = calculateReceiveBTC(amount, currentPrice);
-    setReceiveBTC(btcAmount);
-  }, [amount, price, limitPrice, marketPrice, priceLabel, calculateReceiveBTC]);
-
-  useEffect(() => {
-    const currentPrice = priceLabel === "Price" ? marketPrice : limitPrice;
-    const usdAmount = calculateReceiveUSD(sellAmount, currentPrice);
-    setReceiveUSD(usdAmount);
-  }, [
-    sellAmount,
-    price,
-    limitPrice,
-    marketPrice,
-    priceLabel,
-    calculateReceiveUSD,
-  ]);
-
-  useEffect(() => {
-    if (priceLabel === "Price" && !isInputFocused && marketPrice) {
-      const formattedMarketPrice = formatToTwoDecimalsWithComma(marketPrice);
-      setPrice(formattedMarketPrice);
-      setLimitPrice(formattedMarketPrice);
-    }
-  }, [marketPrice, priceLabel, isInputFocused, formatToTwoDecimalsWithComma]);
-
-  useEffect(() => {
-    if (amount && !isAmountFocused) {
-      const numericValue = amount.replace(/,/g, "");
-      const num = parseFloat(numericValue);
-      const availableBalance = getAvailableBalance();
-      const isValid = !isNaN(num) && num <= availableBalance;
-      setIsAmountValid(isValid);
-
-      if (isValid) {
-        const sliderPercentage = calculateSliderPercentage(numericValue);
-        setSliderValue(sliderPercentage);
-      }
-    }
-  }, [amount, isAmountFocused, getAvailableBalance, calculateSliderPercentage]);
-
-  // Derive UI-specific props
-  const getButtonColor = (type: "buy" | "sell") =>
-    type === "buy"
-      ? "bg-[#309C7D] hover:bg-[#28886C]"
-      : "bg-[#D84C4C] hover:bg-[#C73E3E]";
-
-  const getAmountIcon = (type: "buy" | "sell") =>
-    type === "buy"
-      ? "/currency-icons/dollar-icon.svg"
-      : "/currency-icons/bitcoin-icon.svg";
-
-  const getReceiveIcon = (type: "buy" | "sell") =>
-    type === "buy"
-      ? "/currency-icons/bitcoin-icon.svg"
-      : "/currency-icons/dollar-icon.svg";
-
   return (
     <div>
       <Tabs defaultValue="buy">
@@ -556,68 +26,12 @@ export default function MarketOrderContainer() {
 
         {/* Buy Tab */}
         <TabsContent value="buy" className="mt-7">
-          <OrderForm
-            type="buy"
-            inputRef={inputRef}
-            amountInputRef={amountInputRef}
-            priceLabel={priceLabel}
-            price={price}
-            amount={amount}
-            receiveAmount={receiveBTC}
-            sliderValue={sliderValue}
-            isAmountValid={isAmountValid}
-            isInputFocused={isInputFocused}
-            isAmountFocused={isAmountFocused}
-            availableBalance={formatAvailableBalance()}
-            balanceCurrency="USD"
-            symbol="BTC"
-            buttonColor={getButtonColor("buy")}
-            amountIcon={getAmountIcon("buy")}
-            receiveIcon={getReceiveIcon("buy")}
-            isSubmitting={createBuyOrderMutation.isPending}
-            onPriceFocus={handlePriceFocus}
-            onPriceChange={handlePriceChange}
-            onPriceBlur={handlePriceBlur}
-            onAmountChange={handleAmountChange}
-            onAmountFocus={handleAmountFocus}
-            onAmountBlur={handleAmountBlur}
-            onSliderChange={handleSliderChange}
-            onMarketClick={handleMarketClick}
-            onSubmit={() => handleSubmit("buy")}
-          />
+          <BuyOrderContainer />
         </TabsContent>
 
         {/* Sell Tab */}
         <TabsContent value="sell" className="mt-7">
-          <OrderForm
-            type="sell"
-            inputRef={inputRef}
-            amountInputRef={amountInputRef}
-            priceLabel={priceLabel}
-            price={price}
-            amount={sellAmount}
-            receiveAmount={receiveUSD}
-            sliderValue={sellSliderValue}
-            isAmountValid={isSellAmountValid}
-            isInputFocused={isInputFocused}
-            isAmountFocused={isSellAmountFocused}
-            availableBalance={formatAvailableBTCBalance()}
-            balanceCurrency="BTC"
-            symbol="BTC"
-            buttonColor={getButtonColor("sell")}
-            amountIcon={getAmountIcon("sell")}
-            receiveIcon={getReceiveIcon("sell")}
-            isSubmitting={createSellOrderMutation.isPending}
-            onPriceFocus={handlePriceFocus}
-            onPriceChange={handlePriceChange}
-            onPriceBlur={handlePriceBlur}
-            onAmountChange={handleSellAmountChange}
-            onAmountFocus={handleSellAmountFocus}
-            onAmountBlur={handleSellAmountBlur}
-            onSliderChange={handleSellSliderChange}
-            onMarketClick={handleMarketClick}
-            onSubmit={() => handleSubmit("sell")}
-          />
+          <SellOrderContainer />
         </TabsContent>
       </Tabs>
     </div>
